41 pointsAndy_McKenzie 01 April 2012 10:10:38PM Permalink A few years into this book, I was diagnosed as diabetic and received a questionnaire in the mail. The insurance carrier stated that diabetics often suffer from depression and it was worried about me. One of the questions was “Do you think about death?” Yes, I do. “How often?” the company wanted to know. “Yearly? Monthly? Weekly? Daily?” And if daily, how many times per day? I dutifully wrote in, “About 70 times per day.” The next time I saw my internist, she told me the insurer had recommended psychotherapy for my severe depression. I explained to her why I thought about death all day—merely an occupational hazard—and she suggested getting therapy nonetheless. I thought, fine, it might help with the research.
The therapist found me tragically undepressed, and I asked her if she could help me design a new life that would maximize the few years that I had left. After all, one should have a different life strategy at sixty than at twenty. She asked why I thought I was going to die and why I had such a great fear of death. I said, I am going to die. It’s not a fear; it’s a reality. There must be some behavior that could be contraindicated for a man my age but other normally dangerous behavior that takes advantage of the fact that I am risking fewer years at sixty or sixty-five years of age than I was at twenty or twenty-five (such as crimes that carry a life sentence, crushing at age twenty but less so at age sixty-five). Surely psychology must have something to say on the topic. Turns out, according to my therapist, it does not. There was therapy for those with terminal illness, for the bereaved, for the about-to-be-bereaved, for professionals who dealt with terminal patients, and so on, but there was nothing for people who were simply aware that their life would come to a natural end. It would seem to me that this is a large, untapped market. The therapist advised me not to think about death.

Dick Teresi, The Undead